Add files via upload
This file contains pin mappings for the Anycubic i3 Mega with Ultrabase from 2020. (This config may work on an Anycubic i3 Mega v1 prior to the Ultrabase if you comment out the definition of the endstop_pin in the stepper_z1 section.) To use this config, the firmware should be compiled for the AVR atmega2560. IMPORTANT: This file is  configured for the TMC2208 drivers on X & Y axis and original A4988 on the extruder It also is set for a BTT Dual Geared extruder